add Landmark & Hole denoamalizer + move to correct namespace
