+++ src/Serializer/Normalizer/UserNormalizer.php
@@ -13,9 +13,6 @@ use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 class UserNormalizer implements NormalizerInterface
 {
 
-    /**
-     * {@inheritDoc}
-     */
     public function normalize($user, $format = null, array $context = [])
     {
         return [
@@ -25,9 +22,6 @@ class UserNormalizer implements NormalizerInterface
         ];
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function supportsNormalization($data, $format = null): bool
     {
         return $data instanceof User;
