--- src/Service/ApiTokenGenerator.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Service;
-
-class ApiTokenGenerator
-{
-    private string $prefix = 'mygstatapi-token:';
-
-    public function generate() : string
-    {
-        return uniqid($this->prefix, true);
-    }
-}
